consolidate token/nft send functions between extension/mobile (#3637)
* consolidate nft sending

* consolidate sending transactions
