+++ packages/app-mobile/src/components/BottomDrawerSolanaConfirmation.tsx
@@ -1,24 +1,13 @@
-import type { Connection } from "@solana/web3.js";
 import type { BigNumber } from "ethers";
 
-import { useState } from "react";
-import { View, Text } from "react-native";
+import { Text } from "react-native";
 
-import { programs, tryGetAccount } from "@cardinal/token-manager";
+import { Blockchain, walletAddressDisplay } from "@coral-xyz/common";
 import {
-  Blockchain,
-  confirmTransaction,
-  SOL_NATIVE_MINT,
-  Solana,
-  walletAddressDisplay,
-  metadataAddress,
-} from "@coral-xyz/common";
-import { useSolanaCtx } from "@coral-xyz/recoil";
-import {
-  Metadata,
-  TokenStandard,
-} from "@metaplex-foundation/mpl-token-metadata";
-import { PublicKey } from "@solana/web3.js";
+  useSolanaCtx,
+  useSolanaTransaction,
+  SolTransactionStep,
+} from "@coral-xyz/recoil";
 
 import {
   Error,
@@ -29,8 +18,6 @@ import {
 import { Margin, PrimaryButton, TokenAmountHeader } from "~components/index";
 import { useTheme } from "~hooks/useTheme";
 import { SettingsList } from "~screens/Unlocked/Settings/components/SettingsMenuList";
-
-type Step = "confirm" | "sending" | "complete" | "error";
 
 export function SendSolanaConfirmationCard({
   navigation,
@@ -49,94 +36,16 @@ export function SendSolanaConfirmationCard({
   };
   destinationAddress: string;
   amount: BigNumber;
-  onCompleteStep?: (step: Step) => void;
+  onCompleteStep?: (step: SolTransactionStep) => void;
 }): JSX.Element {
-  const [txSignature, setTxSignature] = useState<string | null>(null);
-  const solanaCtx = useSolanaCtx();
-  const [error, setError] = useState(
-    "Error 422. Transaction time out. Runtime error. Reticulating splines."
-  );
-  const [cardType, setCardType] = useState<Step>("confirm");
-
-  const handleChangeStep = (step: Step) => {
-    setCardType(step);
-    if (onCompleteStep) {
-      onCompleteStep(step);
-    }
-  };
-
-  const onConfirm = async () => {
-    handleChangeStep("sending");
-    //
-    // Send the tx.
-    //
-    let txSig;
-    try {
-      if (token.mint === SOL_NATIVE_MINT.toString()) {
-        txSig = await Solana.transferSol(solanaCtx, {
-          source: solanaCtx.walletPublicKey,
-          destination: new PublicKey(destinationAddress),
-          amount: amount.toNumber(),
-        });
-      } else if (
-        await isCardinalWrappedToken(
-          solanaCtx.connection,
-          token.mint?.toString() as string
-        )
-      ) {
-        txSig = await Solana.transferCardinalToken(solanaCtx, {
-          destination: new PublicKey(destinationAddress),
-          mint: new PublicKey(token.mint!),
-          amount: amount.toNumber(),
-          decimals: token.decimals,
-        });
-      } else if (
-        await isProgrammableNftToken(
-          solanaCtx.connection,
-          token.mint?.toString() as string
-        )
-      ) {
-        txSig = await Solana.transferProgrammableNft(solanaCtx, {
-          destination: new PublicKey(destinationAddress),
-          mint: new PublicKey(token.mint!),
-          amount: amount.toNumber(),
-          decimals: token.decimals,
-          source: new PublicKey(token.address),
-        });
-      } else {
-        txSig = await Solana.transferToken(solanaCtx, {
-          destination: new PublicKey(destinationAddress),
-          mint: new PublicKey(token.mint!),
-          amount: amount.toNumber(),
-          decimals: token.decimals,
-        });
-      }
-    } catch (err: any) {
-      setError(err.toString());
-      handleChangeStep("error");
-      return;
-    }
-
-    setTxSignature(txSig);
-
-    //
-    // Confirm the tx.
-    //
-    try {
-      await confirmTransaction(
-        solanaCtx.connection,
-        txSig,
-        solanaCtx.commitment !== "confirmed" &&
-          solanaCtx.commitment !== "finalized"
-          ? "confirmed"
-          : solanaCtx.commitment
-      );
-      handleChangeStep("complete");
-    } catch (err: any) {
-      setError(err.toString());
-      handleChangeStep("error");
-    }
-  };
+  const { txSignature, onConfirm, cardType, error } = useSolanaTransaction({
+    token,
+    destinationAddress,
+    amount,
+    onComplete: () => {
+      onCompleteStep?.(cardType);
+    },
+  });
 
   return (
     <>
@@ -206,7 +115,7 @@ export function ConfirmSendSolana({
   );
 }
 
-const ConfirmSendSolanaTable: React.FC<{
+export const ConfirmSendSolanaTable: React.FC<{
   destinationAddress: string;
 }> = ({ destinationAddress }) => {
   const theme = useTheme();
@@ -241,49 +150,4 @@ const ConfirmSendSolanaTable: React.FC<{
       menuItems={menuItems}
     />
   );
-};
-
-// TODO(peter) share between mobile/extension
-const isCardinalWrappedToken = async (
-  connection: Connection,
-  tokenAddress: string
-) => {
-  const [tokenManagerId] =
-    await programs.tokenManager.pda.findTokenManagerAddress(
-      new PublicKey(tokenAddress)
-    );
-  const tokenManagerData = await tryGetAccount(() =>
-    programs.tokenManager.accounts.getTokenManager(connection, tokenManagerId)
-  );
-  if (tokenManagerData?.parsed && tokenManagerData?.parsed.transferAuthority) {
-    try {
-      programs.transferAuthority.accounts.getTransferAuthority(
-        connection,
-        tokenManagerData?.parsed.transferAuthority
-      );
-      return true;
-    } catch (error) {
-      console.error(error);
-      console.log("Invalid transfer authority");
-    }
-  }
-  return false;
-};
-
-const isProgrammableNftToken = async (
-  connection: Connection,
-  mintAddress: string
-) => {
-  try {
-    const metadata = await Metadata.fromAccountAddress(
-      connection,
-      await metadataAddress(new PublicKey(mintAddress))
-    );
-
-    return metadata.tokenStandard == TokenStandard.ProgrammableNonFungible;
-  } catch (error) {
-    // most likely this happens if the metadata account does not exist
-    console.log(error);
-    return false;
-  }
 };
